requirements.txt
transformers>=4.35.0
datasets>=2.18.0
python-dotenv>=1.0.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
accelerate>=0.25.0
